PurrfectBytesWeb/pyproject.toml
[project]
name = "purrfectbytes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gtts>=2.5.4",
    "fastapi>=0.100.0",
    "uvicorn>=0.30.0",
    "python-multipart>=0.0.6",
    "jinja2>=3.1.0",
    "moviepy>=1.0.3",
    "pillow>=10.0.0",
    "imageio-ffmpeg>=0.4.9",
    "speechrecognition>=3.10.0",
    "pydub>=0.25.1",
    "librosa>=0.10.0",
    "langdetect>=1.0.9",
    "edge-tts>=6.1.0",
    "google-genai>=1.0.0",
    "openai>=1.0.0",
    "anthropic>=0.30.0",
    "google-api-python-client>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-config",
    "--strict-markers",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.1",
    "pytest-mock>=3.14.1",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.24.0",
    "pytest-mock>=3.11.0",
    "pyfakefs>=5.2.0",
]

dev = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]
